+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This change log follows the conventions of
 
 ## [Unreleased][unreleased]
 
+### Added
+
+- The ability to proxy metadata from mounted archive files corresponding to USB media mounted in the Opus Quad, which cannot itself provide that information.
+
 ### Fixed
 
 - It turns out there is a [bug](https://bugs.java.com/bugdatabase/view_bug?bug_id=8023649)
@@ -26,6 +30,7 @@ This change log follows the conventions of
 - Upgraded to version 0.2.0 of Crate Digger, which had a couple of backwards-incompatible
   changes to accommodate a new linter in the latest release of the Kaitai Struct Compiler
   which identified places where field names were not following the style guide.
+- Upgraded to building Java 11 class files (since the new version of Crate Digger does), so we can take advantage of new APIs.
 
 ## [7.3.0] - 2023-11-24